dic 12

+++ app/src/main/java/com/boxico/android/kn/qrupkeeper/dtos/DataCenter.java
@@ -8,7 +8,7 @@ import java.io.Serializable;
 public class DataCenter implements Serializable{
     @SerializedName("id")
     @Expose
-    private long id = -1;
+    private int id = -1;
 
     @SerializedName("name")
     @Expose
@@ -25,11 +25,11 @@ public class DataCenter implements Serializable{
     private String code;
 
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
